Add accuracy function
Provide a function in Tagger class to compute the accuracy. To achieve
this, an utility module was created and a simple flatten function was
implemented. This function is able to reduce one level of nested lists.
The accuracy can be computed by the weighted average of each sentence
accuracy by using sum, map and lambda.